requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# PDF processing
pdfplumber>=0.10.0

# AI & ML
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
chromadb>=0.4.0

# File monitoring
watchdog>=3.0.0

# Web scraping (if needed)
beautifulsoup4>=4.12.0

# HTTP client
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.2
